+++ SearchEngine.module.php
@@ -30,7 +30,7 @@ use SearchEngine\Config,
  * @method string renderScripts(array $args = []) Render script tags for a given theme.
  * @method string render(array $what = [], array $args = []) Render entire search feature, or optionally just some parts of it (styles, scripts, form, results.)
  *
- * @version 0.10.1
+ * @version 0.11.0
  * @author Teppo Koivula <teppo.koivula@gmail.com>
  * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
  */
@@ -72,6 +72,7 @@ class SearchEngine extends WireData implements Module, ConfigurableModule {
             'FieldtypeRepeater',
             'FieldtypeRepeaterMatrix',
             'FieldtypePageTable',
+            'FieldtypePage',
         ],
         'prefixes' => [
             'link' => 'link:',