+++ lib/Indexer.php
@@ -5,11 +5,18 @@ namespace SearchEngine;
 /**
  * SearchEngine Indexer
  *
- * @version 0.3.1
+ * @version 0.4.0
  * @author Teppo Koivula <teppo.koivula@gmail.com>
  * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
  */
 class Indexer extends Base {
+
+    /**
+     * Meta prefix used for storing (and later identifying) non-field-values
+     *
+     * @var string
+     */
+    const META_PREFIX = '_meta.';
 
     /**
      * Index multiple pages
@@ -74,9 +81,10 @@ class Indexer extends Base {
      * @param \ProcessWire\Page $page
      * @param array $indexed_fields
      * @param string $prefix
+     * @param array $args Additional arguments.
      * @return array
      */
-    protected function ___getPageIndex(\ProcessWire\Page $page, array $indexed_fields = [], string $prefix = ''): array {
+    protected function ___getPageIndex(\ProcessWire\Page $page, array $indexed_fields = [], string $prefix = '', array $args = []): array {
         $index = [];
         if ($page->id && !empty($indexed_fields)) {
             $repeatable_fieldtypes = [
@@ -90,10 +98,21 @@ class Indexer extends Base {
                         // Note: union operator is slightly faster than array_merge() and makes sense
                         // here since we're working with associative arrays only.
                         $index += $this->getRepeatableIndexValue($page, $field, $indexed_fields, $prefix);
+                    } else if ($field->type instanceof \ProcessWire\FieldtypePage) {
+                        $indexed_page_reference_fields = [
+                            'name',
+                            'title',
+                        ];
+                        $index += $this->getPageReferenceIndexValue($page, $field, $indexed_page_reference_fields, $prefix);
                     } else {
                         $index[$prefix . $field->name] = $this->getIndexValue($page, $field);
                     }
                 }
+            }
+            // Check for Page properties, which are not included in the fields property of a Page.
+            if (in_array('name', $indexed_fields)) {
+                $name_prefix = $args['name_prefix'] ?? '';
+                $index[self::META_PREFIX . $prefix . 'name'] = $name_prefix . $this->getIndexValue($page, 'name');
             }
         }
         return $index;
@@ -103,21 +122,25 @@ class Indexer extends Base {
      * Get the index value for a single page field
      *
      * @param \ProcessWire\Page $page
-     * @param \ProcessWire\Field $field
+     * @param \ProcessWire\Field|string $field Field object or name of a field (string).
      * @return mixed
      */
-    protected function ___getIndexValue(\ProcessWire\Page $page, \ProcessWire\Field $field) {
+    protected function ___getIndexValue(\ProcessWire\Page $page, $field) {
         $value = '';
-        if ($field->type instanceof \ProcessWire\FieldtypeFile) {
-            $value = $page->getUnformatted($field->name)->implode(' ', function($item) {
-                return implode(' ', array_filter([
-                    $item->name,
-                    $item->description,
-                ]));
-            });
-        } else {
-            $value = $page->getFormatted($field->name);
+        $field_name = $field;
+        if ($field instanceof \ProcessWire\Field) {
+            if ($field->type instanceof \ProcessWire\FieldtypeFile) {
+                $value = $page->getUnformatted($field->name)->implode(' ', function($item) {
+                    return implode(' ', array_filter([
+                        $item->name,
+                        $item->description,
+                    ]));
+                });
+            } else {
+                $field_name = $field->name;
+            }
         }
+        $value = $page->getFormatted($field_name);
         return $value;
     }
 
@@ -135,7 +158,30 @@ class Indexer extends Base {
         foreach ($page->get($field->name) as $child) {
             // Note: union operator is slightly faster than array_merge() and makes sense
             // here since we're working with associative arrays only.
-            $index += $this->getPageIndex($child, $indexed_fields, $prefix . $child->name . '.');
+            $index += $this->getPageIndex($child, $indexed_fields, $prefix . $field->name . '.' . $child->name . '.');
+        }
+        return $index;
+    }
+
+    /**
+     * Get index value for a Page Reference field
+     *
+     * @param \ProcessWire\Page $page
+     * @param \ProcessWire\Field $field
+     * @param array $indexed_fields
+     * @param string $prefix
+     * @return array
+     */
+    protected function ___getPageReferenceIndexValue(\ProcessWire\Page $page, \ProcessWire\Field $field, array $indexed_fields = [], string $prefix = ''): array {
+        $index = [];
+        $page_ref = $page->getUnformatted($field->name);
+        if ($page_ref instanceof \ProcessWire\PageArray && $page_ref->count()) {
+            $name_prefix = $field->name . ':';
+            foreach ($page_ref as $page_ref_page) {
+                $index += $this->getPageIndex($page_ref_page, $indexed_fields, $prefix . $field->name . '.' . $page_ref_page->name . '.', [
+                    'name_prefix' => $name_prefix,
+                ]);
+            }
         }
         return $index;
     }
@@ -152,10 +198,25 @@ class Indexer extends Base {
     protected function processIndex(array $index): string {
         $processed_index = '';
         if (!empty($index)) {
+            $meta_index = [];
+            foreach ($index as $index_key => $index_value) {
+                // Identify and capture values belonging to the meta index (non-field values).
+                if (strpos($index_key, self::META_PREFIX) === 0) {
+                    $meta_key = substr($index_key, strlen(self::META_PREFIX));
+                    $meta_index[$meta_key] = $index_value;
+                    unset($index[$index_key]);
+                }
+            }
             $processed_index = implode(' ', $index);
-            $processed_index .= ' ' . $this->getURLIndex($processed_index);
+            $url_index = $this->getURLIndex($processed_index);
+            if (!empty($url_index)) {
+                $meta_index['urls'] = $url_index;
+            }
             $processed_index = strip_tags($processed_index);
             $processed_index = preg_replace('/\s+/', ' ', $processed_index);
+            if (!empty($meta_index)) {
+                $processed_index .= json_encode($meta_index, JSON_UNESCAPED_UNICODE);
+            }
         }
         return $processed_index;
     }