+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for indexing Page Reference fields.
+- Support for indexing non-field Page properties (currently applies to name only).
+- New hookable method Indexer::___getPageReferenceIndexValue().
+
 ## [0.10.1] - 2019-09-04
 
 ### Fixed
@@ -37,7 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.8.2] - 2019-07-29
 
 ### Changed
-- Change SearchEngine.info.json format as an attempt to make the Modules directory read it properly.
+- Changed SearchEngine.info.json format as an attempt to make the Modules directory read it properly.
 
 ## [0.8.1] - 2019-07-28